Add files via upload
As with any of my Python programs, use them at your own risk. Create a test folder with copies of your photos and make sure they do what you want before you commit all your precious photos to a program a random non-programmer posted on the internet.

ImageSorter - It will sort you photos by looking at the metadata and them putting them into folders by Year>Month>Day. ie. If the metadata shows you took the photo April 18th 2023. it will be placed in folder '2023', in a subfolder '4', in a subfolder '18'

ImageMetadata - Allows you to add metadata to a photo that will allow you to search from the folder search bar. ie. if you point it to a folder with photos, it will show you the photo and you can enter metadata of what the  photo is. If the photo is of a boat at sunset . You would enter 'boat' 'sunset' (no quotes) and save-next it. Then when you have set the metadat to all your photos, if you do a serach for 'boat' that picture would show up.